refactor: move status type to enums
