+++ backend/app/models/enums.py
@@ -1,0 +1,9 @@
+import enum
+
+
+class StatusType(str, enum.Enum):
+    SCHEDULED = "scheduled"
+    LIVE = "live"
+    FINISHED = "finished"
+    POSTPONED = "postponed"
+    CANCELLED = "cancelled"
