+++ backend/tests/integration/test_matches_repo.py
@@ -3,7 +3,8 @@ from datetime import date, datetime, timezone
 import pytest
 
 from app.api.v1.repositories import matches as matches_repo
-from app.models.match import Match, StageType, StatusType
+from app.models.enums import StatusType
+from app.models.match import Match, StageType
 from app.models.team import Team, TeamType
 from app.models.tournament import Tournament
